chore: split contexts
